Show calculated unit price when pallet details are entered
Update the product management form to display the calculated unit price based on pallet price and units per pallet using parseFloat for accurate conversions.

+++ client/src/pages/product-management.tsx
@@ -1978,7 +1978,7 @@ export default function ProductManagement() {
                               const watchedPalletPrice = form.watch("palletPrice");
                               const watchedUnitsPerPallet = form.watch("unitsPerPallet");
                               const palletPriceNum = parseFloat(watchedPalletPrice);
-                              const unitsPerPalletNum = parseInt(watchedUnitsPerPallet);
+                              const unitsPerPalletNum = parseFloat(watchedUnitsPerPallet);
                               const unitPrice =
                                 palletPriceNum > 0 && unitsPerPalletNum > 0
                                   ? palletPriceNum / unitsPerPalletNum
